Organize Security feature folder
